requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
pandas==1.5.3
pdfminer.six==20221105
pdfplumber==0.8.0
PyYAML==6.0
typing_extensions==4.4.0
